cleaned up mqtt-otel API, reducing instrumentation code needed in mqtt-broker and mqttd
