+++ mqtt/mqtt-otel/Cargo.toml
@@ -14,3 +14,4 @@ prometheus = { version = "0.10.0", features = [ "push" ] }
 hyper = "0.13"
 futures = "0.3"
 tokio = { version = "0.2", features = [ "time" ] }
+lazy_static = "1.4"
